move ReportDetailPage component and all its sub-components to a new directory
