+++ frontend/src/pages/ReportDetail/ReportDetailPage.scss
