+++ frontend/src/pages/ReportDetail/ReportDetailPage.tsx
