+++ frontend/src/pages/ReportDetail/components/ActionButtons.tsx
