+++ frontend/src/pages/ReportDetail/components/AiAnalysisTab.tsx
