+++ frontend/src/pages/ReportDetail/components/AiAssistantNotice.scss
