+++ frontend/src/pages/ReportDetail/components/AiAssistantNotice.tsx
