+++ frontend/src/pages/ReportDetail/components/EmergencyAlert.tsx
