+++ frontend/src/pages/ReportDetail/components/FlaggedValuesSection.tsx
