+++ frontend/src/pages/ReportDetail/components/InfoCard.tsx
