+++ frontend/src/pages/ReportDetail/components/LabValueItem.tsx
