+++ frontend/src/pages/ReportDetail/components/LowConfidenceNotice.tsx
