+++ frontend/src/pages/ReportDetail/components/NormalValuesSection.tsx
