+++ frontend/src/pages/ReportDetail/components/OriginalReportTab.tsx
