+++ frontend/src/pages/ReportDetail/components/ReportHeader.tsx
