+++ frontend/src/pages/ReportDetail/components/ReportTabs.tsx
